requirements.txt
scikit-learn==0.21.3
matplotlib==3.0.3
numpy==1.17.3
pandas==0.25.3
tensorflow==2.1.0
IPython==7.4.0
graphviz==0.10.1
torch
faiss-gpu
